Reorganize generator further into its own subcommand
