+++ src/generator/data_types/graph.rs
